Refactored login to the MVP model
