+++ app/src/main/java/com/appTec/RegistrateApp/presenter/LoginPresenter.java
@@ -3,6 +3,8 @@ package com.appTec.RegistrateApp.presenter;
 import android.app.Activity;
 import android.content.Context;
 
+import com.appTec.RegistrateApp.models.UserCredential;
+
 public interface LoginPresenter {
     /**
      * Interface for login presenter
@@ -12,8 +14,14 @@ public interface LoginPresenter {
 
     void verifyPreviousLogin(); // To the interactor
     void navigateToNextView();  // To the view
-    void handleLogin(String email, String password); // To the interactor
+    void handleLogin(UserCredential userCredential); // To the interactor
     void handleFirstRun(Activity activity); // When is the first time tha the app run on the device.
     boolean isTheFirstRun();
+
+    void showLoginProgressDialog(String message);
+    void hideLoginProgressDialog();
+
+    void showMessage(String title, String message);
+    void showAlertDialog(String title, String message);
 
 }
