+++ app/src/main/java/com/appTec/RegistrateApp/util/Constants.java
@@ -17,16 +17,7 @@ public class Constants {
 
     // Shared preferences key words values
     public static final String SHARED_PREFERENCES_GLOBAL = "com.appTec.RegistrateApp.PREFERENCES_FILE_KEY"; // Storage
-    public static final String CURRENT_STATE = "CURRENT_STATE";                                     //
-    public static final String LOGIN_USER_STATE = "LOGGED";                                         //
 
-    public static final String STATE_WORKING = "WORKING";                                           //
-    public static final String STATE_NOT_WORKING = "NOT WORKING";                                   //
-
-    public static final String LOGGED_USER = "LOGGED";                                              //
-    public static final String NOT_LOGGED_USER = "NOT LOGGED";                                      //
-
-    public static final String USER_TOKEN = "USERTOKEN";                                            //
 
     public static final String LAST_EXIT_TIME = "LAST_EXIT_TIME";                                   //
 
@@ -34,6 +25,8 @@ public class Constants {
     public static final String USER_ACCESS_TOKEN = "ACCESS_TOKEN";                                  // Storage a string
     public static final String USER_REFRESH_TOKEN = "REFRESH_TOKEN";                                // Storage a string
     public static final String CURRENT_IMEI = "CURRENT_IMEI";                                       // Storage a string
+    public static final String IS_USER_WORKING = "IS_WORKING";                                      // Storage a boolean
+    public static final String IS_USER_LOGGED = "IS_LOGGED";                                        // Storage a boolean
 
 
 }
