+++ app/src/main/java/com/appTec/RegistrateApp/view/LoginActivity.java
@@ -12,7 +12,6 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
-import android.os.Build;
 import android.os.Bundle;
 import android.telephony.TelephonyManager;
 import android.text.TextUtils;
@@ -54,29 +53,25 @@ import retrofit2.Response;
 
 public class LoginActivity extends Activity implements View.OnClickListener, LoginActivityView {
 
-    //UI elements
+    // UI elements
     private ImageButton btnLogin;
     private EditText txtEmail;
     private EditText txtPassword;
     ProgressDialog progressDialog;
 
-    //Business logic elements
+    // Business logic elements
     private TelephonyManager telephonyManager;
 
-    /*
-    =======================================
-    ACTIVITY LIFECYCLE METHODS
-    =======================================
-     */
+
 
     // Instance the presenter
     LoginPresenterImpl loginPresenter = new LoginPresenterImpl(this);
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
-        setTheme(R.style.SplashTheme);
+        setTheme(R.style.SplashTheme);                                                  // Showing the splash screen for until the activity is ready
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.login_activity);
+        setContentView(R.layout.login_activity);                                        // Binding the layout
 
         // Binding UI elements
         txtEmail = (EditText) findViewById(R.id.email);
@@ -123,7 +118,7 @@ public class LoginActivity extends Activity implements View.OnClickListener, Log
                     if (checkSelfPermission(permission.READ_PHONE_STATE) == PackageManager.PERMISSION_GRANTED) {
                         UserCredential userCredential = new UserCredential(email, password);
                         showLoginProgressDialog(getString(R.string.login_progress_dialog_message));
-                        login(userCredential);
+                        loginPresenter.handleLogin(userCredential);
                     } else {
                         ActivityCompat.requestPermissions(this, new String[]{android.Manifest.permission.READ_PHONE_STATE}, 225);
                     }
@@ -133,143 +128,7 @@ public class LoginActivity extends Activity implements View.OnClickListener, Log
 
     }
 
-    //Login user
-    public void login(final UserCredential userCredential) {
-        LoginRetrofitInterface loginRetrofitInterface = ApiClient.getClient().create(LoginRetrofitInterface.class);
-        Call<JsonObject> call = loginRetrofitInterface.login(userCredential);
-        call.enqueue(new Callback<JsonObject>() {
-            @Override
-            public void onResponse(Call<JsonObject> call, Response<JsonObject> response) {
-                if (response.code() == 200) {
-                    if (android.os.Build.VERSION.SDK_INT >= 23) {
-                        if (checkSelfPermission(Manifest.permission.READ_PHONE_STATE) != PackageManager.PERMISSION_GRANTED) {
-                            ActivityCompat.requestPermissions(LoginActivity.this, new String[]{android.Manifest.permission.READ_PHONE_STATE}, 225);
-                        } else {
-                            if (android.os.Build.VERSION.SDK_INT >= 23 && android.os.Build.VERSION.SDK_INT < 26) {
-                                String deviceImei = telephonyManager.getDeviceId();
-                            }
-                            if (android.os.Build.VERSION.SDK_INT >= 26) {
-                                String deviceImei = telephonyManager.getImei();
-                            }
-                        }
-                    }
-                    User user = new User();
-                    Company company = new Company();
-                    ArrayList<WorkingPeriod> workingPeriodList = new ArrayList<WorkingPeriod>();
-                    user.setId(response.body().getAsJsonObject("data").get("id").getAsInt());
-                    user.setName(response.body().getAsJsonObject("data").get("nombres").getAsString());
-                    user.setLastName(response.body().getAsJsonObject("data").get("apellidos").getAsString());
-                    user.setEmail(response.body().getAsJsonObject("data").get("email").getAsString());
-                    company.setName(response.body().getAsJsonObject("data").getAsJsonObject("empresa").get("nombre").getAsString());
-                    company.setLatitude(response.body().getAsJsonObject("data").getAsJsonObject("empresa").get("latitud").getAsDouble());
-                    company.setLongitude(response.body().getAsJsonObject("data").getAsJsonObject("empresa").get("longitud").getAsDouble());
-                    company.setRadius(response.body().getAsJsonObject("data").getAsJsonObject("empresa").get("radio").getAsFloat());
-                    user.setCompany(company);
-                    ApiClient.setToken(response.body().get("token").toString().replace("\"", ""));
-                    DatabaseAdapter.getDatabaseAdapterInstance().insertUser(user);
-                    DatabaseAdapter.getDatabaseAdapterInstance().insertCompany(company);
-                    StaticData.setCurrentUser(user);
-                    StaticData.getCurrentUser().setCompany(company);
-                    changeWorkingState(Constants.STATE_NOT_WORKING);
-                    setLoggedUser();
-                    findUserDevice();
-                } else if (response.code() == 404 || response.code() == 401) {
-                    hideLoginProgressDialog();
-                    showCredentialsErrorMessage();
-                } else if (response.code() == 502) {
-                    hideLoginProgressDialog();
-                } else {
-                    hideLoginProgressDialog();
-                    try {
-                        JSONObject errorJson = new JSONObject(response.errorBody().string());
-                        showErrorMessage(errorJson.toString());
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
 
-            @Override
-            public void onFailure(Call<JsonObject> call, Throwable t) {
-                showConectionErrorMessage();
-            }
-
-        });
-    }
-
-    //Get user device
-    public void findUserDevice() {
-        DeviceRetrofitInterface deviceRetrofitInterface = ApiClient.getClient().create(DeviceRetrofitInterface.class);
-        Log.d("User ", StaticData.getCurrentUser().getId()+"");
-        Call<JsonObject> deviceCall = deviceRetrofitInterface.get(ApiClient.getToken(), StaticData.getCurrentUser().getId());
-        Log.d("Tokeen", ApiClient.getToken());
-        deviceCall.enqueue(new Callback<JsonObject>() {
-            @Override
-            public void onResponse(Call<JsonObject> call, Response<JsonObject> response) {
-                Log.d("deviceImei", "DEVICE ON RESPONSE OK");
-
-
-                JsonArray deviceList = response.body().getAsJsonArray("data");
-                Log.d("REEEEESSSPOONSE", deviceList.size()+"");
-                boolean deviceFound = false;
-                for (int i = 0; i < deviceList.size() && deviceFound == false; i++) {
-                    JsonObject jsonDevice = deviceList.get(i).getAsJsonObject();
-
-                    if (SharedPreferencesHelper.getSharedPreferencesInstance().getString(Constants.CURRENT_IMEI, "").equals(jsonDevice.get("imei").getAsString())) {
-                        Device device = new Device();
-                        int deviceId = jsonDevice.get("id").getAsInt();
-                        String deviceName = jsonDevice.get("nombre").getAsString();
-                        String deviceModel = jsonDevice.get("modelo").getAsString();
-                        String deviceImei = jsonDevice.get("imei").getAsString();
-                        boolean deviceStatus = jsonDevice.get("estado").getAsBoolean();
-                        device.setId(deviceId);
-                        device.setName(deviceName);
-                        device.setModel(deviceModel);
-                        device.setImei(deviceImei);
-                        device.setStatus(deviceStatus);
-                        DatabaseAdapter.getDatabaseAdapterInstance().insertDevice(device);
-                        deviceFound = true;
-                    }
-                }
-                findPermissionTypes();
-            }
-
-            @Override
-            public void onFailure(Call<JsonObject> call, Throwable t) {
-
-            }
-        });
-
-    }
-
-    //Get PermissionType
-    public void findPermissionTypes() {
-        PermissionTypeRetrofitInterface permissionTypeRetrofitInterface = ApiClient.getClient().create(PermissionTypeRetrofitInterface.class);
-        final Call<JsonObject> permissionTypeCall = permissionTypeRetrofitInterface.get(ApiClient.getToken());
-        permissionTypeCall.enqueue(new Callback<JsonObject>() {
-            @Override
-            public void onResponse(Call<JsonObject> call, Response<JsonObject> response) {
-                JsonArray jsonLstPermissionType = response.body().getAsJsonArray("data");
-                if (jsonLstPermissionType.size() > 0) {
-                    for (int i = 0; i < jsonLstPermissionType.size(); i++) {
-                        JsonObject jsonPermissionType = jsonLstPermissionType.get(i).getAsJsonObject();
-                        int permissionTypeId = jsonPermissionType.get("id").getAsInt();
-                        String permissionTypeName = jsonPermissionType.get("nombre").getAsString();
-                        PermissionType permissionType = new PermissionType(permissionTypeId, permissionTypeName);
-                        DatabaseAdapter.getDatabaseAdapterInstance().insertPermissionType(permissionType);
-                        StaticData.getPermissionTypes().add(permissionType);
-                    }
-                }
-                hideLoginProgressDialog();
-                navigateToNextView();
-            }
-
-            @Override
-            public void onFailure(Call<JsonObject> call, Throwable t) {
-
-            }
-        });
-    }
 
     //Change to BottomNavigation activity
     public void navigateToNextView() {
@@ -291,23 +150,6 @@ public class LoginActivity extends Activity implements View.OnClickListener, Log
 
 
 
-    private void setLoggedUser() {
-        SharedPreferences sharedPref = this.getSharedPreferences(
-                Constants.SHARED_PREFERENCES_GLOBAL, Context.MODE_PRIVATE);
-        SharedPreferences.Editor editor = sharedPref.edit();
-        editor.putString(Constants.LOGIN_USER_STATE, Constants.LOGGED_USER);
-        editor.commit();
-    }
-
-
-
-    private void changeWorkingState(String state) {
-        SharedPreferences sharedPref = this.getSharedPreferences(
-                Constants.SHARED_PREFERENCES_GLOBAL, Context.MODE_PRIVATE);
-        SharedPreferences.Editor editor = sharedPref.edit();
-        editor.putString(Constants.CURRENT_STATE, state);
-        editor.commit();
-    }
 
     //Dialogs
     public void showLoginProgressDialog(String message) {
@@ -321,19 +163,16 @@ public class LoginActivity extends Activity implements View.OnClickListener, Log
         progressDialog.dismiss();
     }
 
-    public void showErrorMessage(String error) {
-        showDialog("Error", error);
+
+    @Override
+    public void showMessage(String title, String message) {
+        showAlertDialog(title, message);
     }
 
-    public void showCredentialsErrorMessage() {
-        showDialog("Autenticación fallida", "El usuario y contraseña proporcionados no son correctos.");
-    }
 
-    public void showConectionErrorMessage() {
-        showDialog("Error de conexión", "Al parecer no hay conexión a Internet.");
-    }
 
-    public void showDialog(String title, String message) {
+    @Override
+    public void showAlertDialog(String title, String message) {
         new AlertDialog.Builder(this)
                 .setTitle(title)
                 .setMessage(message)
